Added Interim Fix for NHibernate Config
I'm effectively just changing the location of where the config is being
searched, but we should expose an application interface that caches
config information and can be injected into handler constructors, rather
that IApplicationEnvironment. Otherwise, we should look at refactoring
the way we create the session so that it has access to config itself.

Fixes #20